Move away from using Array as the default child type
Tag children work with the Inner type in the UI library to represent zero, one, or many items.

The Inner type is monadic and has lots of useful functionality that allows for nested functions, dom elements, or whatever the dev wants.

There's a concept of NoInner also, which is an Inner Unit - that stops children being added at all and will throw a compile time exception